% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakout.hist.R
\name{breakout_hist}
\alias{breakout_hist}
\title{Plots a faceted histogram with multiple variable frequancies.}
\usage{
breakout_hist()
}
\value{
plot A faceted out muli-variable histogram color coded with a legend
}
\description{
Plots a faceted histogram with multiple variable frequancies.
}
