% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carnivorous_tree.R
\name{carnivorous_tree}
\alias{carnivorous_tree}
\title{Plots a phylogenetic tree of species}
\usage{
species<-c("Pinguicula lutea","Sarracenia alata", "Drosera tracyi","Pinguicula pumila")
carnivorous_tree(species)
}
\arguments{
\item{Vspec}{A vector of species names}
}
\value{
product A plotted phylogenetic tree of those species
}
\description{
Plots a phylogenetic tree of species
}
