% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col.merge.R
\name{col.merge}
\alias{col.merge}
\title{merges two columns into one.}
\usage{
col.merge(gbutterflies,Scientific_name,Genus,Species)
}
\arguments{
\item{newname}{The new name of the merged column}

\item{col1}{The first column you want to merge.}

\item{col2}{The second column you want to merge.}

\item{Dataset}{A dataset}
}
\value{
product A dataset like the input data set but with the two columns merged.
}
\description{
merges two columns into one.
}
