% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hutcheson_t.R
\name{hutcheson_t}
\alias{hutcheson_t}
\title{Performs a hutcheson-T test for two samples of an abundance dataset}
\usage{
hutcheson_t(grouped_data$'1',grouped_data$'3')
}
\arguments{
\item{sample1}{Vector of abundance for a sample}

\item{sample2}{Vector of abundance for another sample}
}
\description{
Performs a hutcheson-T test for two samples of an abundance dataset
}
