% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shan_hutch_prep.R
\name{shan_hutch_prep}
\alias{shan_hutch_prep}
\title{turns a dataset into an abundance dataframe for Shannons index and Hutcheson-t test}
\usage{
shan_hutch_prep(gbutterflies,SITE,Species)
}
\arguments{
\item{dataset}{a dataset}

\item{samplecol}{A column of samples you want to index}

\item{speciescol}{A column of species}
}
\value{
A dataset of abundances
}
\description{
turns a dataset into an abundance dataframe for Shannons index and Hutcheson-t test
}
