% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannons_index.R
\name{shannons_index}
\alias{shannons_index}
\title{plots a dataset of shannons indexes}
\usage{
shannons_index(grouped_data)
}
\arguments{
\item{dataset}{A dataframe of abundances}
}
\value{
A plot of different shannons indexes
}
\description{
plots a dataset of shannons indexes
}
